QUESTIONS_GENERATION/requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
cmake==3.27.2
colorama==0.4.6
comm==0.1.4
dataclasses-json==0.5.14
datasets==2.14.4
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.101.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gradio_client==0.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.4.0
langchain==0.0.264
langsmith==0.0.22
line-profiler==4.1.0
lit==16.0.6
markdownify==0.11.6
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.0
numexpr==2.8.5
numpy==1.24.1
oauthlib==3.2.2
openapi-schema-pydantic==1.2.4
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.1
Pygments==2.16.1
PyMuPDF==1.22.5
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.2
scipy==1.11.2
selenium==4.14.0
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data==0.6.3
starlette==0.27.0
sympy==1.11.1
tenacity==8.2.3
tokenizers==0.13.3
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
transformers==4.31.0
trio==0.22.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
undetected-chromedriver==3.5.3
uritemplate==4.1.1
urllib3==1.26.13
uvicorn==0.23.2
wcwidth==0.2.8
websockets==11.0.3
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
